+++ src/app/ar-assets/_components/ArAssetCard.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import Link from 'next/link';
 import { ARAsset } from '../_types';
 import { Button } from '@/shared/components/common/Button';
